RainbowLED(): Initialize the RGB LED pattern in a const struct.
This eliminates runtime overhead. (TODO: Check if it actually had
runtime overhead before, or if the compiler optimized it.)

Re-added the return value. Note that previously, the function still had
a return type of void, but it returned a value anyway, and -fpermissive
allowed it.

ptmsysmSetInfoLedPattern(): Take RGBLedPattern by pointer, not by value.

Size difference:
   text    data     bss     dec     hex filename
  59624     132    1244   61000    ee48 build/main.o [before]
  58936     136    1244   60316    eb9c build/main.o [after]
   -688      +4       0    -684    -2ac difference