Reworked the dialog functions to take a text parameter.
Renamed all "dialogue" variables and functions to "dialog". "dialogue"
refers to speech, while "dialog" refers to a pop-up window.

dialoguetext has been removed in favor of setting the test when calling
DialogBoxAppear() and DialogBoxDisappear(). These functions will store
the last text assigned in an std::string, so if called with nullptr,
the previous text will be used.

Renamed dialoguebox.png to dialogbox.png.