Handle the "fully charged and plugged in" case.
PTM:GetAdapterState can be used to determine if the AC adapter is
plugged in. 3dbrew incorrectly states that it's the same as
PTM:GetBatteryChargeState, which is probably why ctrulib doesn't
have a wrapper for it.

It turns out that PTM:GetAdapterState always returns whether the
AC adapter is connected or not, whereas PTM:GetBatteryChargeState
returns 1 only while charging; once it's finished charging, it
returns 0.

ptmu_x.c: Added a wrapper for PTM:GetAdapterState. Note that we can't
access the ptm:u service handle from ctrulib, so we have to open our
own service handle.

Added some missing service exit calls to ensure everything is cleaned up.

References:
- https://3dbrew.org/wiki/PTM:GetAdapterState
- https://github.com/smealum/ctrulib/blob/master/libctru/source/services/ptmu.c