Allow starting game cards that have unreadable banners.
This may fix support for blocked flashcards, e.g. CycloDS.

NOTE: I'm not sure if FSUSER_CardSlotIsInserted() returns 1
for blocked flashcards, so this will need some testing.

Also, if the card is inserted but type is CARD_TYPE_UNKNOWN, the card
will be continuously polled until it gets a valid type. This is needed
because some cards don't respond immediately.